+++ bazel_tidy_example/helper.cpp
@@ -1,0 +1,10 @@
+#include "helper.hpp"
+
+A* func_return_pointer(){
+  return new A;
+}
+
+int helper(){
+  //int b = 10;
+  return 0;
+}