+++ bazel_tidy_example/helper.hpp
@@ -1,0 +1,10 @@
+
+class A{
+
+};
+
+int helper();
+
+
+A* func_return_pointer();
+