+++ bazel_tidy_example/main2.cpp
@@ -1,0 +1,13 @@
+class A{};
+
+A* addTwo(int num) 
+{ 
+    return new A; 
+} 
+
+using B= int;
+
+double complex_function(A& a, B& b) 
+{ 
+  return 20; 
+}