+++ bazel_tidy_example/return_test.cpp
@@ -1,0 +1,7 @@
+class A{};
+
+A* ReturnTest(int a){
+  if(a>0)
+    return new A;
+  return nullptr;
+}